Enable mode IMEX solver choices in the driver
Previously the IMEX solver in the driver only supported
`split_explicit_implicit = false`, namely the use of two full models and
not the remainder model. Now is the IMEX solver is configured with
`split_explicit_implicit = true` the remainder model will be used.

An additional boolean parameter `discrete_splitting` is also added which
allows the user to specify whether they want to have a discrete
splitting or PDE level splitting done in the remainder model.